authorization implemented

+++ src/api/notes/handler.js
@@ -18,7 +18,9 @@ class NotesHandler {
 
       const { title = 'untitle', body, tags } = request.payload;
 
-      const noteId = await this._service.addNote({ title, body, tags });
+      const { id: credentialId } = request.auth.credentials;
+
+      const noteId = await this._service.addNote({ title, body, tags, owner: credentialId });
 
       const response = h.response({
         status: 'success',
@@ -56,8 +58,10 @@ class NotesHandler {
     }
   }
 
-  async getNotesHandler() {
-    const notes = await this._service.getNotes();
+  async getNotesHandler(request) {
+    const { id: credintialId } = request.auth.credentials;
+
+    const notes = await this._service.getNotes(credintialId);
 
     return {
       status: 'success',
@@ -70,7 +74,9 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credintialId } = request.auth.credentials;
 
+      await this._service.verifyNoteOwner(id, credintialId);
       const note = await this._service.getNoteById(id);
 
       const response = h.response({
@@ -114,6 +120,10 @@ class NotesHandler {
 
       const { id } = request.params;
 
+      const { id: credintialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credintialId);
+
       await this._service.editNoteById(id, request.payload);
 
       return {
@@ -148,6 +158,10 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+
+      const { id: credintialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credintialId);
 
       await this._service.deleteNoteById(id, request.payload);
